refactor: [Review Requests] refactoring
