+++ src/components/review-requests-table/index.js
@@ -1,3 +1,3 @@
-import ReviewRequestsTableCreation from './review-requests-table-creation';
+import ReviewRequestsTableCreation from './review-requests-table';
 
 export default ReviewRequestsTableCreation;
