+++ src/store/reducers/index.js
@@ -7,9 +7,9 @@ import tasks from './tasks/tasks';
 import tasksTableData from './tasks/tasks-table';
 import reviews from './reviews';
 import login from './login';
-import crossCheckSessionsData from './cross-check-sessions-data';
-import crossCheckSessions from './cross-check-sessions';
-import crossCheckSession from './cross-check-session';
+import crossCheckSessionsData from './cross-check-sessions/cross-check-sessions-data';
+import crossCheckSessions from './cross-check-sessions/cross-check-sessions';
+import crossCheckSession from './cross-check-sessions/cross-check-session';
 import reviewRequestsData from './review-requests/review-requests-data';
 import reviewRequest from './review-requests/review-request';
 import values from './values';
